Enable sprites of DMG while on CGB mode.
Use undocumented register 0xff6c least significant byte to check if game
is DMG or CGB. If DMG, use `bank0`. Otherwise, use `bank1`.

We already do this check for other cases, probably just forgot for gfx.

Undocumented register `0xff6c`:
the LSB in 0xff6c is both readable and writable in CGB. It is set by
the boot ROM: 1 for DMG and 0 for CGB.