Do not force set APU to AGB mode with declicking.
By setting DMG/CGB to use AGB-mode apu with declicking, we fail one of
the tests from `Demotronic` Demo. More especifically, the sound one
(expects `0xff`, but reads a random value).

The reason we used to set AGB-mode with declicking is due to extra
processing for declicking in this mode.