[GBA] Disable part of SoundDriverVsyncOff()
The official BIOS executes a copy from the stack pointer data to another
area in the internal working RAM. This is problematic to emulate because
our emulated BIOS does not touch the stack. In addition, the official
BIOS overwrites the first word in the stack with "0". While this is fine
with an official BIOS (the stack points to some internal BIOS data at
this point), this results in a user stack corruption with an emulated
BIOS.

Reproducing the BIOS behavior when doing this copy is near impossible
since the stack and CPU states are very different with the emulated
BIOS. Furthermore, the reason why this copy is done by the official BIOS
is unknown. The most likely explanation is that this a leftover from a
debug attempt and is not actually required for most intents and purposes
so this patch disables the copy.

This fixes Shrek games on trunk.