Fix joysticks with SDL video
For joysticks and game controllers to work with SDL video output modes,
the joystick background input hint has to be set for SDL3. No idea what
happens with SDL2.

Also fix an issue Claude found with an incorrect reference to the XRC
element for our own option of that name, which this change does not
affect because we check the focus.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>