Dialogs: Save dialog DPI
In case dialogs are restored on a screen with a different DPI setting,
they may end up having the wrong size. This fixes the issue by checking
the DPI when restoring the dialog size and position.
In addition, this fixes an issue in wxWidgets where dialogs could be
shown out of the screen bounds if the screen setup has changed since the
last time they were displayed.