build: Pentium3 arch for XP with -mfpmath=sse
Build the Windows XP build with `-march=pentium3 -mmmx -msse
-mfpmath=sse`. Previously the options were incorrect and SSE2 was being
used, which Pentium 3 does not support.

Turn off `-ffast-math` for the Windows XP build in favor of
`-mfpmath=sse` because the floating point precision is insufficient.

Add `-fexpensive-optimizations` to the list of optimization flags.

Remove `-mtune=generic`, because that breaks some of the other options.

The suggestion about SSE2 was from Claude.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>