More fixes for the display config dialog
Refactor the basic tab in the XRC for the display config dialog and add
a couple of `Fit()` calls to the C++.

Also fix the on screen display tab.

This fixes the corruption when SDL is selected on XP.

Some US government shits are still breaking the layout on the 64 bit
binary.

Fixed with the help of Claude.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>