RTC Improvements
So the cause of Pokemon Gen 2 Time not Set on new vbam battery saves was because the overflow bit was being set immediately at initializing the RTC due to the initial value of Last Time being 0 (Epoch, 1970) and then immediately being brought up to the current time via time_t

we now set Last Time to -1, and initialize immediately to Now instead of 0.

Also has the initial implementation of the rtc halt bit, though it doesn't seem to pass tests so might not actually be working.

we also log the RTC to console in debug builds

And finally, a save check is added to reset stuck overflow control values on already affected save files.