made a mistake with preserving bit 7
it is software writable, and properly allowing this does allow a new battery to be created without the epoch fixes, but does not correct batteries with low days, when control stuck on 128.

we'll do this to allow the games to overflow and reset the bit by themselves as necessary, as intended, but also maintain the other change where we don't set epoch 0 on the initial data.